fix(relay): tighten unpaired token observation
