+++ ARCHITECTURE.md
@@ -536,7 +536,6 @@ src/
 │   │   │   ├── redaction.rs
 │   │   │   ├── relay_auto_heal.rs
 │   │   │   ├── relay_dead_reattach.rs
-│   │   │   ├── relay_health_snapshot.rs
 │   │   │   ├── relay_progress.rs
 │   │   │   ├── runtime_resolve.rs
 │   │   │   ├── session_enrichment.rs
