+++ src/services/discord/health.rs
@@ -30,7 +30,6 @@ mod recovery;
 mod redaction;
 mod relay_auto_heal;
 mod relay_dead_reattach;
-mod relay_health_snapshot;
 mod relay_progress;
 mod runtime_resolve;
 mod session_enrichment;
