+++ src/services/discord/health/snapshot.rs
@@ -5,11 +5,11 @@ use super::liveness_authority::CaptureCoordinateObservation;
 use super::mailbox::MailboxHealthSnapshot;
 use super::provider_probe::{self, ProviderHealthSnapshot};
 use super::redaction;
-use super::relay_health_snapshot::{RelayHealthBuildInput, build_relay_health_snapshot};
 use super::session_enrichment::SessionEnrichment;
 use super::stall_verdict;
 use super::transcript_binding_stall::{self, resolve_bound_selector};
 use super::unpaired_active_token;
+use super::unpaired_active_token::{RelayHealthBuildInput, build_relay_health_snapshot};
 use super::{BotTokenReloadScopes, HealthRegistry, bot_token_reload_scopes};
 use crate::services::discord;
 use crate::services::discord::SharedData;
