+++ src/services/discord/relay_health.rs
@@ -157,19 +157,6 @@ impl RelayHealthSnapshot {
                 .is_some_and(|capture| capture > self.last_relay_offset)
             && self.unread_bytes.is_some_and(|bytes| bytes > 0)
     }
-
-    fn has_fresh_relay_progress_for_active_turn(&self) -> bool {
-        matches!(
-            (
-                self.mailbox_turn_started_at_ms,
-                self.last_relay_ts_ms,
-                self.last_relay_age_secs,
-            ),
-            (Some(turn_started_at), Some(relay_at), Some(relay_age))
-                if relay_at >= turn_started_at
-                    && relay_age < UNPAIRED_ACTIVE_TOKEN_PROGRESS_FRESH_SECS
-        )
-    }
 }
 
 /// Time allowed for a newly minted mailbox token to acquire its durable
@@ -177,11 +164,6 @@ impl RelayHealthSnapshot {
 /// Its initial value happens to equal the stall-watchdog threshold, but the
 /// two policies have different meanings and no reason to move together.
 pub(in crate::services::discord) const UNPAIRED_ACTIVE_TOKEN_GRACE_SECS: u64 = 600;
-
-/// A confirmed relay heartbeat this recent proves that a rowless turn is
-/// still delivering. This policy is intentionally independent of both the
-/// mint-to-row grace above and the stall-watchdog threshold.
-const UNPAIRED_ACTIVE_TOKEN_PROGRESS_FRESH_SECS: u64 = 600;
 
 pub(in crate::services::discord) fn observation_age_secs(
     observed_at_ms: i64,
@@ -226,7 +208,6 @@ impl RelayStallClassifier {
             && snapshot
                 .mailbox_turn_age_secs
                 .is_some_and(|age| age >= UNPAIRED_ACTIVE_TOKEN_GRACE_SECS)
-            && !snapshot.has_fresh_relay_progress_for_active_turn()
         {
             return RelayStallState::UnpairedActiveToken;
         }
@@ -355,6 +336,21 @@ mod tests {
                 RelayStallState::QueueBlocked,
             ),
             (
+                "young rowless active token remains foreground before grace",
+                RelayHealthSnapshot {
+                    active_turn: RelayActiveTurn::Foreground,
+                    tmux_alive: Some(true),
+                    watcher_attached: true,
+                    mailbox_has_cancel_token: true,
+                    mailbox_active_user_msg_id: Some(9001),
+                    mailbox_turn_started_at_ms: Some(1_000_000),
+                    mailbox_turn_age_secs: Some(599),
+                    unpaired_active_token_reconfirmed: true,
+                    ..RelayHealthSnapshot::test_snapshot()
+                },
+                RelayStallState::ActiveForegroundStream,
+            ),
+            (
                 "old rowless active token with null relay coordinates is unpaired",
                 RelayHealthSnapshot {
                     active_turn: RelayActiveTurn::Foreground,
@@ -370,7 +366,7 @@ mod tests {
                 RelayStallState::UnpairedActiveToken,
             ),
             (
-                "rowless active token with fresh relay progress stays active",
+                "channel relay telemetry does not exempt an old rowless active token",
                 RelayHealthSnapshot {
                     active_turn: RelayActiveTurn::Foreground,
                     tmux_alive: Some(true),
@@ -381,55 +377,6 @@ mod tests {
                     last_relay_ts_ms: Some(1_600_000),
                     last_relay_age_secs: Some(1),
                     last_relay_offset: 0,
-                    unpaired_active_token_reconfirmed: true,
-                    ..RelayHealthSnapshot::test_snapshot()
-                },
-                RelayStallState::ActiveForegroundStream,
-            ),
-            (
-                "fresh relay coordinate from a previous turn does not exempt",
-                RelayHealthSnapshot {
-                    active_turn: RelayActiveTurn::Foreground,
-                    tmux_alive: Some(true),
-                    watcher_attached: true,
-                    mailbox_has_cancel_token: true,
-                    mailbox_turn_started_at_ms: Some(2_000_000),
-                    mailbox_turn_age_secs: Some(1_200),
-                    last_relay_ts_ms: Some(1_999_999),
-                    last_relay_age_secs: Some(1),
-                    last_relay_offset: 128,
-                    unpaired_active_token_reconfirmed: true,
-                    ..RelayHealthSnapshot::test_snapshot()
-                },
-                RelayStallState::UnpairedActiveToken,
-            ),
-            (
-                "positive relay offset without a timestamp does not exempt",
-                RelayHealthSnapshot {
-                    active_turn: RelayActiveTurn::Foreground,
-                    tmux_alive: Some(true),
-                    watcher_attached: true,
-                    mailbox_has_cancel_token: true,
-                    mailbox_turn_started_at_ms: Some(1_000_000),
-                    mailbox_turn_age_secs: Some(1_200),
-                    last_relay_offset: 128,
-                    unpaired_active_token_reconfirmed: true,
-                    ..RelayHealthSnapshot::test_snapshot()
-                },
-                RelayStallState::UnpairedActiveToken,
-            ),
-            (
-                "stale relay timestamp from this turn does not exempt",
-                RelayHealthSnapshot {
-                    active_turn: RelayActiveTurn::Foreground,
-                    tmux_alive: Some(true),
-                    watcher_attached: true,
-                    mailbox_has_cancel_token: true,
-                    mailbox_turn_started_at_ms: Some(1_000_000),
-                    mailbox_turn_age_secs: Some(1_200),
-                    last_relay_ts_ms: Some(1_600_000),
-                    last_relay_age_secs: Some(600),
-                    last_relay_offset: 128,
                     unpaired_active_token_reconfirmed: true,
                     ..RelayHealthSnapshot::test_snapshot()
                 },
